Synchronize with Polygon repository

+++ problems/mayday/sources/wrong_answer.cpp
@@ -4,10 +4,10 @@ using namespace std;
 #define SIZE 200001
 
 int n;
-long long vx[SIZE];
-long long vy[SIZE];
-multiset<long long> x_prod_v;
-multiset<long long> y_prod_v;
+int vx[SIZE];
+int vy[SIZE];
+multiset<int> x_prod_v;
+multiset<int> y_prod_v;
 
 int ans = 0;
 
@@ -16,8 +16,8 @@ int main() {
     for (int i=1; i<=n; i++) cin >> vx[i];
     for (int i=1; i<=n; i++) cin >> vy[i];
 
-    for (long long i=1; i<=n; i++) x_prod_v.insert(i * vx[i]);
-    for (long long i=1; i<=n; i++) y_prod_v.insert(i * vy[i]);
+    for (int i=1; i<=n; i++) x_prod_v.insert(i * vx[i]);
+    for (int i=1; i<=n; i++) y_prod_v.insert(i * vy[i]);
 
     for (auto it=x_prod_v.begin(); it!=x_prod_v.end(); it++)
         if (y_prod_v.find(*it) != y_prod_v.end()) {
